(extension) convert PanelStateBroadcaster from renderless component to plain module
Replaces the React renderless-component pattern (returning null, using hooks)
with a plain useStore.subscribe() subscription started at controller boot in
index.tsx. The broadcaster logic is unchanged; only the React scaffolding is
removed.
